Make backfillContractDiffs a CLI option for backfillStorage
